+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - **[Breaking change]** Update dependencies. Use `esm` instead of `@std/esm`, update Typescript to `2.8.3`.
+- **[Fix]** Fix Node processes spawn on Windows (Mocha, Nyc)
 
 ## 0.16.2 (2017-02-07)